feat: allow users to specify a subject line in their contact emails
A subject field is now shown in the contact form, allowing users to specify a custom subject instead of the default one. This creates an order of priority for subject line contents - first user defined, then config defined, then the default defined in the build_subject function signature. The logic goes through each of these options until a valid value is found and then this is used as the subject for the email. The email field in the contact form is completely optional and no validation is done on it.